chore(tests): move account test to common/wallet-core
