+++ suite-common/test-utils/src/configureMockStore.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/ban-types */
 import { configureStore } from '@reduxjs/toolkit';
 import {
     Action,
