+++ suite-common/wallet-core/tsconfig.json
@@ -6,6 +6,7 @@
         { "path": "../suite-config" },
         { "path": "../suite-types" },
         { "path": "../suite-utils" },
+        { "path": "../test-utils" },
         { "path": "../wallet-config" },
         { "path": "../wallet-types" },
         { "path": "../wallet-utils" }
